fix(groups): restore owned-groups listing on /groups (#180)
* fix(groups): show owned groups again on /groups tab

`getBaseAndCmgGroupsByOwnerBatch` was rewritten during the SDK swap to call
`sdk.rpc.group.getGroupMemberships()`, which returns groups the avatar is a
member of — not groups it owns. Owners who created groups but never joined
them as members had those groups vanish from the "My groups" tab and the
Safe-switcher dropdown.

Replace with a `sdk.rpc.group.getGroups({ ownerIn, groupTypeIn })` paginated
query against `V_CrcV2.Groups`. The alternative SDK helper
`sdk.rpc.group.findGroups()` accepts an `ownerIn` argument but the underlying
`circles_findGroups` RPC method does not honor it server-side, so it returns
an empty list — verified against the production indexer.

Also delete the orphaned duplicate at
`src/lib/areas/groups/utils/getGroupsByOwnerBatch.ts`, which used the broken
`findGroups()` path and was not imported anywhere.

All three call sites (`/groups`, `/register`, `safeDiscovery`) already
import from `$lib/shared/utils/getGroupsByOwnerBatch`, so no caller changes
are required.

* docs(safeDiscovery): tighten membership-merge comment

The previous comment claimed the owner-filter "misses base groups whose
owner field is null". Probing the production indexer shows every
CrcV2_BaseGroupCreated (360) and CrcV2_CMGroupCreated (137) row has a
non-null owner; only the legacy CrcV2_RegisterGroup type (98 rows) has
null owners and is therefore unmatched by the owner-filter. Reword the
comment to reflect that and to also call out the more common reason for
the merge: avatars that are members of groups they don't own.
